Remove MutationEngine.
