--- lib/MutationEngine.cpp
@@ -1,18 +1,0 @@
-#include "MutationEngine.h"
-
-#include "MutationPoint.h"
-#include "llvm/IR/Constants.h"
-#include "llvm/IR/Instruction.h"
-#include "llvm/IR/InstrTypes.h"
-#include "llvm/IR/Module.h"
-
-using namespace llvm;
-using namespace mull;
-
-void MutationEngine::applyMutation(llvm::Module *M, MutationPoint &MP) {
-  MP.applyMutation(M);
-}
-
-void MutationEngine::revertMutation(MutationPoint &MP) {
-  //MP.revertMutation();
-}
